+++ Elements/src/GeometricElement.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Reflection;
 using System.Runtime.CompilerServices;
 using Elements.Geometry;
 using Elements.Geometry.Solids;
@@ -26,9 +25,15 @@ namespace Elements
         internal Csg.Solid _csg;
 
         /// <summary>
+        /// The element's animation.
+        /// </summary>
+        [JsonIgnore]
+        public Animation Animation { get; set; }
+
+        /// <summary>
         /// The element's bounds.
         /// The bounds are only available when the geometry has been
-        /// updated using UpdateBoundsAndComputeSolid(),
+        /// updated using UpdateBoundsAndComputeSolid().
         /// </summary>
         [JsonIgnore]
         public BBox3 Bounds => _bounds;