+++ Elements/src/Serialization/glTF/GltfExtensions.cs
@@ -16,6 +16,8 @@ using System.Net;
 using SixLabors.ImageSharp.Formats.Png;
 using SixLabors.ImageSharp;
 using Image = glTFLoader.Schema.Image;
+using Vector3 = Elements.Geometry.Vector3;
+using Quaternion = Elements.Geometry.Quaternion;
 using System.Reflection;
 
 [assembly: InternalsVisibleTo("Hypar.Elements.Tests")]
@@ -498,7 +500,14 @@ namespace Elements.Serialization.glTF
             }
         }
 
-        private static int AddAccessor(List<Accessor> accessors, int bufferView, int byteOffset, Accessor.ComponentTypeEnum componentType, int count, float[] min, float[] max, Accessor.TypeEnum accessorType)
+        private static int AddAccessor(List<Accessor> accessors,
+                                       int bufferView,
+                                       int byteOffset,
+                                       Accessor.ComponentTypeEnum componentType,
+                                       int count,
+                                       float[] min,
+                                       float[] max,
+                                       Accessor.TypeEnum accessorType)
         {
             var a = new Accessor
             {
@@ -991,6 +1000,7 @@ namespace Elements.Serialization.glTF
             var textures = new List<Texture>();
             var images = new List<Image>();
             var samplers = new List<Sampler>();
+            var animations = new List<glTFLoader.Schema.Animation>();
             var materials = gltf.Materials != null ? gltf.Materials.ToList() : new List<glTFLoader.Schema.Material>();
 
             var meshElementMap = new Dictionary<Guid, List<int>>();
@@ -1028,6 +1038,7 @@ namespace Elements.Serialization.glTF
                                             meshTransformMap,
                                             currLines,
                                             drawEdges,
+                                            animations,
                                             mergeVertices);
                 }
                 catch (Exception ex)
@@ -1090,6 +1101,10 @@ namespace Elements.Serialization.glTF
             {
                 gltf.Samplers = samplers.ToArray(samplers.Count);
             }
+            if (animations.Count > 0)
+            {
+                gltf.Animations = animations.ToArray(animations.Count);
+            }
             gltf.Nodes = nodes.ToArray(nodes.Count);
             if (meshes.Count > 0)
             {
@@ -1128,10 +1143,12 @@ namespace Elements.Serialization.glTF
                                                     Dictionary<Guid, Transform> meshTransformMap,
                                                     List<Vector3> lines,
                                                     bool drawEdges,
+                                                    List<glTFLoader.Schema.Animation> animations,
                                                     bool mergeVertices = false)
         {
             var materialId = BuiltInMaterials.Default.Id.ToString();
             int meshId = -1;
+            int nodeId = -1;
 
             if (e is GeometricElement element)
             {
@@ -1189,7 +1206,9 @@ namespace Elements.Serialization.glTF
                                                                 nodes,
                                                                 materialId,
                                                                 ref meshId,
-                                                                content);
+                                                                content,
+                                                                out nodeId,
+                                                                mergeVertices);
                         if (!meshElementMap.ContainsKey(e.Id))
                         {
                             meshElementMap.Add(e.Id, new List<int> { meshId });
@@ -1210,7 +1229,9 @@ namespace Elements.Serialization.glTF
                                                             nodes,
                                                             materialId,
                                                             ref meshId,
-                                                            geometricElement);
+                                                            geometricElement,
+                                                            out nodeId,
+                                                            mergeVertices);
                     if (meshId > -1 && !meshElementMap.ContainsKey(e.Id))
                     {
                         meshElementMap.Add(e.Id, new List<int> { meshId });
@@ -1284,6 +1305,161 @@ namespace Elements.Serialization.glTF
                                      nodes,
                                      ge.Transform);
                 }
+
+                if (ge.Animation != null && nodeId != -1)
+                {
+                    // https://github.com/KhronosGroup/glTF-Tutorials/blob/master/gltfTutorial/gltfTutorial_007_Animations.md
+
+                    var channels = new List<AnimationChannel>();
+                    var animationSamplers = new List<AnimationSampler>();
+
+                    if (ge.Animation.HasAnimatedScale())
+                    {
+                        var scaleTimeBufferView = AddBufferView(bufferViews, 0, buffer.Count, ge.Animation.ScaleTimes.Length, null, null);
+                        buffer.AddRange(ge.Animation.ScaleTimes);
+                        var scaleTimeAccessor = AddAccessor(accessors,
+                                                            scaleTimeBufferView,
+                                                            0,
+                                                            Accessor.ComponentTypeEnum.FLOAT,
+                                                            ge.Animation.ScaleTimes.Length / sizeof(float),
+                                                            new float[] { ge.Animation.ScaleTimeMin },
+                                                            new float[] { ge.Animation.ScaleTimeMax },
+                                                            Accessor.TypeEnum.SCALAR);
+
+                        var scaleBufferView = AddBufferView(bufferViews, 0, buffer.Count, ge.Animation.Scales.Length, null, null);
+                        buffer.AddRange(ge.Animation.Scales);
+                        var scaleAccessor = AddAccessor(accessors,
+                                                            scaleBufferView,
+                                                            0,
+                                                            Accessor.ComponentTypeEnum.FLOAT,
+                                                            ge.Animation.Scales.Length / sizeof(float) / 3,
+                                                            ge.Animation.ScaleMin,
+                                                            ge.Animation.ScaleMax,
+                                                            Accessor.TypeEnum.VEC3);
+
+                        var scaleSampler = new AnimationSampler
+                        {
+                            Input = scaleTimeAccessor,  // time
+                            Output = scaleAccessor,    // scale
+                            Interpolation = AnimationSampler.InterpolationEnum.LINEAR
+                        };
+
+                        var scaleTarget = new AnimationChannelTarget
+                        {
+                            Node = nodeId,
+                            Path = AnimationChannelTarget.PathEnum.scale
+                        };
+
+                        var scaleChannel = new AnimationChannel()
+                        {
+                            Target = scaleTarget,
+                            Sampler = animationSamplers.Count
+                        };
+
+                        animationSamplers.Add(scaleSampler);
+                        channels.Add(scaleChannel);
+                    }
+                    if (ge.Animation.HasAnimatedTranslation())
+                    {
+                        var translationTimeBufferView = AddBufferView(bufferViews, 0, buffer.Count, ge.Animation.TranslationTimes.Length, null, null);
+                        buffer.AddRange(ge.Animation.TranslationTimes);
+                        var translationTimeAccessor = AddAccessor(accessors,
+                                                            translationTimeBufferView,
+                                                            0,
+                                                            Accessor.ComponentTypeEnum.FLOAT,
+                                                            ge.Animation.TranslationTimes.Length / sizeof(float),
+                                                            new float[] { ge.Animation.TranslationTimeMin },
+                                                            new float[] { ge.Animation.TranslationTimeMax },
+                                                            Accessor.TypeEnum.SCALAR);
+
+                        var translationBufferView = AddBufferView(bufferViews, 0, buffer.Count, ge.Animation.Translations.Length, null, null);
+                        buffer.AddRange(ge.Animation.Translations);
+                        var translationAccessor = AddAccessor(accessors,
+                                                            translationBufferView,
+                                                            0,
+                                                            Accessor.ComponentTypeEnum.FLOAT,
+                                                            ge.Animation.Translations.Length / sizeof(float) / 3,
+                                                            ge.Animation.TranslationMin,
+                                                            ge.Animation.TranslationMax,
+                                                            Accessor.TypeEnum.VEC3);
+
+                        var translationSampler = new AnimationSampler
+                        {
+                            Input = translationTimeAccessor,  // time
+                            Output = translationAccessor,    // scale
+                            Interpolation = AnimationSampler.InterpolationEnum.LINEAR
+                        };
+
+                        var translationTarget = new AnimationChannelTarget
+                        {
+                            Node = nodeId,
+                            Path = AnimationChannelTarget.PathEnum.translation
+                        };
+
+                        var translationChannel = new AnimationChannel()
+                        {
+                            Target = translationTarget,
+                            Sampler = animationSamplers.Count
+                        };
+
+                        animationSamplers.Add(translationSampler);
+                        channels.Add(translationChannel);
+                    }
+                    if (ge.Animation.HasAnimatedRotation())
+                    {
+                        var rotationTimeBufferView = AddBufferView(bufferViews, 0, buffer.Count, ge.Animation.RotationTimes.Length, null, null);
+                        buffer.AddRange(ge.Animation.RotationTimes);
+                        var rotationTimeAccessor = AddAccessor(accessors,
+                                                            rotationTimeBufferView,
+                                                            0,
+                                                            Accessor.ComponentTypeEnum.FLOAT,
+                                                            ge.Animation.RotationTimes.Length / sizeof(float),
+                                                            new float[] { ge.Animation.RotationTimeMin },
+                                                            new float[] { ge.Animation.RotationTimeMax },
+                                                            Accessor.TypeEnum.SCALAR);
+
+                        var rotationBufferView = AddBufferView(bufferViews, 0, buffer.Count, ge.Animation.Rotations.Length, null, null);
+                        buffer.AddRange(ge.Animation.Rotations);
+                        var rotationAccessor = AddAccessor(accessors,
+                                                            rotationBufferView,
+                                                            0,
+                                                            Accessor.ComponentTypeEnum.FLOAT,
+                                                            ge.Animation.Rotations.Length / sizeof(float) / 4,
+                                                            ge.Animation.RotationMin,
+                                                            ge.Animation.RotationMax,
+                                                            Accessor.TypeEnum.VEC4);
+
+                        var rotationSampler = new AnimationSampler
+                        {
+                            Input = rotationTimeAccessor,  // time
+                            Output = rotationAccessor,    // scale
+                            Interpolation = AnimationSampler.InterpolationEnum.LINEAR
+                        };
+
+                        var rotationTarget = new AnimationChannelTarget
+                        {
+                            Node = nodeId,
+                            Path = AnimationChannelTarget.PathEnum.rotation
+                        };
+
+                        var rotationChannel = new AnimationChannel()
+                        {
+                            Target = rotationTarget,
+                            Sampler = animationSamplers.Count
+                        };
+
+                        animationSamplers.Add(rotationSampler);
+                        channels.Add(rotationChannel);
+                    }
+
+                    var anim = new glTFLoader.Schema.Animation()
+                    {
+                        Channels = channels.ToArray(),
+                        Samplers = animationSamplers.ToArray()
+                    };
+
+                    animations.Add(anim);
+                }
             }
 
             if (e is ITessellate geo)
@@ -1316,7 +1492,7 @@ namespace Elements.Serialization.glTF
                 var geom = (GeometricElement)e;
                 if (!geom.IsElementDefinition)
                 {
-                    NodeUtilities.CreateNodeForMesh(meshId, nodes, geom.Transform);
+                    nodeId = NodeUtilities.CreateNodeForMesh(meshId, nodes, geom.Transform);
                 }
             }
         }
@@ -1428,10 +1604,14 @@ namespace Elements.Serialization.glTF
                                                            List<Node> nodes,
                                                            string materialId,
                                                            ref int meshId,
-                                                           GeometricElement geometricElement)
+                                                           GeometricElement geometricElement,
+                                                           out int nodeId,
+                                                           bool mergeVertices = false)
         {
             geometricElement.UpdateRepresentations();
             geometricElement.UpdateBoundsAndComputeSolid();
+
+            nodeId = -1;
 
             // TODO: Remove this when we get rid of UpdateRepresentation.
             // The only reason we don't fully exclude openings from processing
@@ -1462,7 +1642,7 @@ namespace Elements.Serialization.glTF
 
                 if (!geometricElement.IsElementDefinition)
                 {
-                    NodeUtilities.CreateNodeForMesh(meshId, nodes, geometricElement.Transform);
+                    nodeId = NodeUtilities.CreateNodeForMesh(meshId, nodes, geometricElement.Transform);
                 }
                 return meshId;
             }