+++ Elements/src/Serialization/glTF/NodeUtilities.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using Elements.Collections.Generics;
 using Elements.Geometry;
 using glTFLoader.Schema;
+using System.Numerics;
+using Vector3 = Elements.Geometry.Vector3;
 
 namespace Elements.Serialization.glTF
 {
@@ -52,14 +54,52 @@ namespace Elements.Serialization.glTF
                 var b = transform.YAxis;
                 var c = transform.ZAxis;
 
-                var transNode = new Node();
-
-                transNode.Matrix = new[]{
+                var transNode = new Node
+                {
+                    Matrix = new[]{
                     (float)a.X, (float)a.Y, (float)a.Z, 0.0f,
                     (float)b.X, (float)b.Y, (float)b.Z, 0.0f,
                     (float)c.X, (float)c.Y, (float)c.Z, 0.0f,
                     (float)transform.Origin.X,(float)transform.Origin.Y,(float)transform.Origin.Z, 1.0f
+                }
                 };
+
+                parentId = AddNode(nodes, transNode, 0);
+            }
+
+            return parentId;
+        }
+
+        internal static int CreateAndAddRTSNode(List<Node> nodes, Transform transform, int parentId)
+        {
+            if (transform != null)
+            {
+                var transNode = new Node();
+
+                // HACK: Using the matrix class from System.Numerics
+                // because it has support for matrix decomposition 
+                // out of the box.
+                var m = new Matrix4x4((float)transform.Matrix.m11,
+                                          (float)transform.Matrix.m12,
+                                          (float)transform.Matrix.m13,
+                                          0.0f,
+                                          (float)transform.Matrix.m21,
+                                          (float)transform.Matrix.m22,
+                                          (float)transform.Matrix.m23,
+                                          0.0f,
+                                          (float)transform.Matrix.m31,
+                                          (float)transform.Matrix.m32,
+                                          (float)transform.Matrix.m33,
+                                          0.0f,
+                                          (float)transform.Origin.X,
+                                          (float)transform.Origin.Y,
+                                          (float)transform.Origin.Z,
+                                          1.0f);
+
+                Matrix4x4.Decompose(m, out var scale, out System.Numerics.Quaternion rotation, out var translation);
+                transNode.Scale = new float[] { scale.X, scale.Y, scale.Z };
+                transNode.Translation = new float[] { translation.X, translation.Y, translation.Z };
+                transNode.Rotation = new float[] { rotation.X, rotation.Y, rotation.Z, rotation.W };
 
                 parentId = AddNode(nodes, transNode, 0);
             }
@@ -83,9 +123,11 @@ namespace Elements.Serialization.glTF
             // transform, so that the transform can be modified in explore at
             // runtime (e.g. by a transform override) and have the expected effect.
             float[] elementTransform = TransformToMatrix(transform);
-            var newNode = new glTFLoader.Schema.Node();
-            newNode.Name = $"{instanceElementId}";
-            newNode.Matrix = elementTransform;
+            var newNode = new glTFLoader.Schema.Node
+            {
+                Name = $"{instanceElementId}",
+                Matrix = elementTransform
+            };
             nodes.Add(newNode);
             newNode.Children = new[] { nodes.Count };
 
@@ -95,8 +137,10 @@ namespace Elements.Serialization.glTF
             // back to Y up further up in the node hierarchy. 
             rootTransform.Rotate(new Vector3(1, 0, 0), 90.0);
             float[] glbOrientationTransform = TransformToMatrix(rootTransform);
-            var elementOrientationNode = new glTFLoader.Schema.Node();
-            elementOrientationNode.Matrix = glbOrientationTransform;
+            var elementOrientationNode = new glTFLoader.Schema.Node
+            {
+                Matrix = glbOrientationTransform
+            };
             nodes.Add(elementOrientationNode);
             elementOrientationNode.Children = new[] { nodes.Count };
 
@@ -107,8 +151,10 @@ namespace Elements.Serialization.glTF
 
         private static int RecursivelyCopyNode(List<Node> nodes, ProtoNode nodeToCopy)
         {
-            var newNode = new Node();
-            newNode.Matrix = nodeToCopy.Matrix;
+            var newNode = new Node
+            {
+                Matrix = nodeToCopy.Matrix
+            };
             if (nodeToCopy.Mesh != null)
             {
                 newNode.Mesh = nodeToCopy.Mesh;
@@ -159,18 +205,21 @@ namespace Elements.Serialization.glTF
         {
             var parentId = 0;
 
-            parentId = NodeUtilities.CreateAndAddTransformNode(nodes, transform, parentId);
+            parentId = CreateAndAddRTSNode(nodes, transform, parentId);
 
             // Add mesh node to gltf nodes
-            var node = new Node();
-            node.Mesh = meshId;
+            var node = new Node
+            {
+                Mesh = meshId,
+            };
+
             var nodeId = AddNode(nodes, node, parentId);
             return nodeId;
         }
 
-        internal static void CreateNodeFromNode(List<glTFLoader.Schema.Node> nodes, Node parentNode, Transform transform)
+        internal static void CreateNodeFromNode(List<glTFLoader.Schema.Node> nodes, Transform transform)
         {
-            var parentId = NodeUtilities.CreateAndAddTransformNode(nodes, transform, 0);
+            CreateAndAddTransformNode(nodes, transform, 0);
         }
     }
 }